Pull out interface for SettingsProvider
